requirements.txt
django==2.2.6
django-cities-light